+++ pendulum/period.py
@@ -186,9 +186,6 @@ class Period(WordableDurationMixin, BaseDuration):
         )
 
     def range(self, unit, amount=1):
-        return list(self.xrange(unit, amount))
-
-    def xrange(self, unit, amount=1):
         method = 'add'
         op = operator.le
         if not self._absolute and self.invert:
@@ -238,7 +235,7 @@ class Period(WordableDurationMixin, BaseDuration):
         return Duration(seconds=self.total_seconds())
 
     def __iter__(self):
-        return self.xrange('days')
+        return self.range('days')
 
     def __contains__(self, item):
         from .datetime import DateTime