+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - `set_to_string_format()/reset_to_string_format()` methods have been removed.
 - The `to_xxx_string()` methods have been removed in favor of a unique `to_string()` method.
 - `from_format()` now uses the alternative formatter tokens.
+- Removed `xrange()` method of the `Period` class and made `range()` a generator.
 
 
 ## [1.2.4] - 2017-06-20